Remve LoginRequestInteractor class
